Megacommit: Actual game engine beginnings
Implemented some parts of my old engine and wrapped SDL around it to do
rendering.
Needs quite a lot of work still, mainly what is blocking the engines
function is the input handing in InitGameState, otherwise a simple tech
demo is in place.

Committing this since another programmer friend joined the project so we
can work on subsystems of the engine simultaneously :)