Committed beginnings of input abstraction
There is no longer a blocking while(true) loop in InitGameState and the
FSM can proceed without obstruction.

There is a lot still to be done so that this can be called a game engine
but the beginnings are starting to take shape.

Replaced lots of printfs with SDL_Logs since SDL eats out all printfs
for funsies... Annoying... Whatever, return to this sometime later and
let printfs do their thing without stupid forced file logging from SDL.

State machine needs work, there is some issue with debug build with the
unordered_map that creates a cryptic error, something wrong with
dereferencing the pointer to the iterator for the first startup, this
needs asap fixing, debug builds are important.